% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_theme.R
\name{app_theme}
\alias{app_theme}
\title{Custom ggplot2 theme}
\usage{
app_theme(
  font_sizes_ = font_sizes,
  font_families_ = font_families,
  size_margin_ = size_margin
)
}
\value{
a ggplot2 theme object
}
\description{
Custom ggplot2 theme
}
